requirements.txt
Django==4.2.16
djangorestframework==3.15.2
celery==5.3.6
redis==5.0.8
django-redis==5.4.0
cryptography==46.0.5
python-dotenv==1.0.1
PyMySQL==1.1.1
gunicorn==22.0.0
# Unit tests
pytest==8.3.2
pytest-django==4.8.0
pluggy==1.6.0
iniconfig==2.3.0
packaging==26.0
colorama==0.4.6
